Move alosi engine subclass to engine/engines.py
